small improvements to my 2 advent of code solutions

+++ src/test/java/lars/advent2018/Day2.java
@@ -13,22 +13,22 @@ public class Day2 {
 
     @Test
     void no_letter_twice() {
-        assertThat(howOften(2L, "abcdef")).isEqualTo(0);
+        assertThat(howOften(2, "abcdef")).isEqualTo(0);
     }
 
     @Test
     void no_letter_triple() {
-        assertThat(howOften(3L, "abcdef")).isEqualTo(0);
+        assertThat(howOften(3, "abcdef")).isEqualTo(0);
     }
 
     @Test
     void one_letter_twice() {
-        assertThat(howOften(2L, "bababc")).isEqualTo(1);
+        assertThat(howOften(2, "bababc")).isEqualTo(1);
     }
 
     @Test
     void one_letter_triple() {
-        assertThat(howOften(3L, "bababc")).isEqualTo(1);
+        assertThat(howOften(3, "bababc")).isEqualTo(1);
     }
 
     @Test
@@ -41,14 +41,14 @@ public class Day2 {
         List<String> lines = input.lines().collect(Collectors.toList());
         var twos = 0;
         for (String l : lines) {
-            int i = howOften(2L, l);
+            int i = howOften(2, l);
             if (i == 1) {
                 twos++;
             }
         }
         var threes = 0;
         for (String l : lines) {
-            int i = howOften(3L, l);
+            int i = howOften(3, l);
             if (i == 1) {
                 threes++;
             }
@@ -56,10 +56,10 @@ public class Day2 {
         return twos * threes;
     }
 
-    private int howOften(long target, String input) {
+    private int howOften(int target, String input) {
         Map<String, Long> map = input.chars()
                 .mapToObj(i -> String.valueOf((char) (i)))
                 .collect(Collectors.groupingBy(Function.identity(), Collectors.counting()));
-        return map.values().contains(target) ? 1 : 0;
+        return map.values().contains((long)target) ? 1 : 0;
     }
 }
